chore(cli/tests): move macos_shared_libraries test to macos_test.rs
